Make step print the right inst in verbose mode.
Since our hooks run before the instruction is executed, we need to
ensure that `step` runs similar to the normal `start` function.
We print the previous instruction right before the next instruction runs
this way we can know and print the state in the comment adjacent to the
instruction. This has relied on the fact that we run a INST hook for
printing the instruction without executing the last instruction when
stopping execution (this is what happens with `emu_stop`).

Using the "count" argument to `emu_start` breaks this, since it doesn't
run the last INST hook, instead stopping before it. If results are
printed to the screen, this causes step to be two instructions behind
instead of just one. Making `step` similar to `start` seems to be a
reasonable solution to this.